- New MyLibrary and MyProfile fragments - Downgrade constraint layout version - NavGraph modified - Modified bottom navigation view style and items
