+++ app/src/main/java/com/android/sergiobelda/gramophone/ui/home/HomeFragment.kt
@@ -1,4 +1,4 @@
-package com.android.sergiobelda.gramophone.ui.main.playlists
+package com.android.sergiobelda.gramophone.ui.home
 
 
 import android.os.Bundle
@@ -12,14 +12,14 @@ import com.android.sergiobelda.gramophone.R
 /**
  * A simple [Fragment] subclass.
  */
-class PlaylistsFragment : Fragment() {
+class HomeFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_playlists, container, false)
+        return inflater.inflate(R.layout.fragment_home, container, false)
     }
