+++ app/src/main/java/com/android/sergiobelda/gramophone/ui/mylibrary/MyLibraryFragment.kt
@@ -1,46 +1,51 @@
-package com.android.sergiobelda.gramophone.ui.main
-
+package com.android.sergiobelda.gramophone.ui.mylibrary
 
 import android.os.Bundle
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentStatePagerAdapter
-
+import androidx.viewpager.widget.ViewPager
 import com.android.sergiobelda.gramophone.R
-import com.android.sergiobelda.gramophone.ui.ErrorFragment
-import com.android.sergiobelda.gramophone.ui.main.pages.*
 import com.android.sergiobelda.gramophone.entities.Album
 import com.android.sergiobelda.gramophone.entities.Song
-import com.android.sergiobelda.gramophone.ui.main.playlists.PlaylistsFragment
-
+import com.android.sergiobelda.gramophone.ui.ErrorFragment
+import com.android.sergiobelda.gramophone.ui.mylibrary.pages.AlbumsFragment
+import com.android.sergiobelda.gramophone.ui.mylibrary.pages.ArtistsFragment
+import com.android.sergiobelda.gramophone.ui.mylibrary.pages.SongsFragment
+import com.android.sergiobelda.gramophone.ui.mylibrary.pages.PlaylistsFragment
+import com.google.android.material.tabs.TabLayout
 
 /**
  * A simple [Fragment] subclass.
- *
+ * Use the [MyLibraryFragment.newInstance] factory method to
+ * create an instance of this fragment.
  */
-class MenuFragment : Fragment() {
+class MyLibraryFragment : Fragment() {
+    private var viewPager : ViewPager? = null
+    private var tabLayout : TabLayout? = null
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_menu, container, false)
+        return inflater.inflate(R.layout.fragment_my_library, container, false)
     }
-
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        viewPager = view.findViewById(R.id.view_pager)
+        tabLayout = view.findViewById(R.id.tab_layout)
         setViewPager()
     }
 
     private fun setViewPager() {
         val pagerAdapter = PagerAdapter(this.fragmentManager!!)
-        //viewPager.adapter = pagerAdapter
-        //tabLayout.setupWithViewPager(viewPager)
+        viewPager?.adapter = pagerAdapter
+        tabLayout?.setupWithViewPager(viewPager)
     }
 
     inner class PagerAdapter(fm : FragmentManager) : FragmentStatePagerAdapter(fm) {
@@ -60,21 +65,8 @@ class MenuFragment : Fragment() {
                     }
                 }
                 2 -> {
-                    fragment = SongsFragment()
-                    fragment.songSelectedListener = object : SongsFragment.SongSelectedListener {
-                        override fun onSongSelected(song: Song) {
-                            //playerBottomSheetBehavior.state = BottomSheetBehavior.STATE_EXPANDED
-                            //playerSongTitleTextView.text = song.title
-                            //playerSongArtistTextView.text = song.artist
-
-                        }
-                    }
-                }
-                3 -> {
-                    fragment = ErrorFragment()
-                }
-                4 -> {
-                    fragment = PlaylistsFragment()
+                    fragment =
+                        PlaylistsFragment()
                 }
                 else -> {
                     fragment = ErrorFragment()
@@ -83,15 +75,13 @@ class MenuFragment : Fragment() {
             return fragment
         }
 
-        override fun getCount(): Int  = 5
+        override fun getCount(): Int = 3
 
         override fun getPageTitle(position: Int): CharSequence? {
             return when (position) {
                 0 -> resources.getString(R.string.artists)
                 1 -> resources.getString(R.string.albums)
-                2 -> resources.getString(R.string.songs)
-                3 -> resources.getString(R.string.categories)
-                4 -> resources.getString(R.string.playlists)
+                2 -> resources.getString(R.string.playlists)
                 else -> null
             }
         }
