+++ app/src/main/java/com/android/sergiobelda/gramophone/ui/mylibrary/pages/ArtistsFragment.kt
@@ -1,25 +1,26 @@
-package com.android.sergiobelda.gramophone.ui.main.home
+package com.android.sergiobelda.gramophone.ui.mylibrary.pages
 
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 
 import com.android.sergiobelda.gramophone.R
 
 /**
  * A simple [Fragment] subclass.
+ *
  */
-class HomeFragment : Fragment() {
+class ArtistsFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home, container, false)
+        return inflater.inflate(R.layout.fragment_artists, container, false)
     }
