+++ app/src/main/java/com/android/sergiobelda/gramophone/ui/mylibrary/pages/PlaylistsFragment.kt
@@ -1,0 +1,26 @@
+package com.android.sergiobelda.gramophone.ui.mylibrary.pages
+
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+
+import com.android.sergiobelda.gramophone.R
+
+/**
+ * A simple [Fragment] subclass.
+ */
+class PlaylistsFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_playlists, container, false)
+    }
+
+
+}
