+++ app/src/main/java/com/android/sergiobelda/gramophone/ui/myprofile/MyProfileFragment.kt
@@ -1,27 +1,24 @@
-package com.android.sergiobelda.gramophone.ui.main.pages
-
+package com.android.sergiobelda.gramophone.ui.myprofile
 
 import android.os.Bundle
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.Fragment
 
 import com.android.sergiobelda.gramophone.R
 
 /**
  * A simple [Fragment] subclass.
- *
  */
-class ArtistsFragment : Fragment() {
+class MyProfileFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_artists, container, false)
+        return inflater.inflate(R.layout.fragment_my_profile, container, false)
     }
-
 
 }
