+++ app/src/main/res/drawable/ic_outline_library_music_24.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+        android:width="24dp"
+        android:height="24dp"
+        android:viewportWidth="24.0"
+        android:viewportHeight="24.0"
+        android:tint="?attr/colorControlNormal">
+    <path
+        android:fillColor="@android:color/white"
+        android:pathData="M20,2L8,2c-1.1,0 -2,0.9 -2,2v12c0,1.1 0.9,2 2,2h12c1.1,0 2,-0.9 2,-2L22,4c0,-1.1 -0.9,-2 -2,-2zM20,16L8,16L8,4h12v12zM12.5,15c1.38,0 2.5,-1.12 2.5,-2.5L15,7h3L18,5h-4v5.51c-0.42,-0.32 -0.93,-0.51 -1.5,-0.51 -1.38,0 -2.5,1.12 -2.5,2.5s1.12,2.5 2.5,2.5zM4,6L2,6v14c0,1.1 0.9,2 2,2h14v-2L4,20L4,6z"/>
+</vector>
