+++ app/src/main/res/layout/albums_list_item.xml
@@ -5,8 +5,10 @@
         app:cardElevation="4dp"
         app:cardCornerRadius="8dp"
         android:clickable="true"
-        android:layout_height="244dp"
-        android:layout_width="172dp"
+        android:layout_height="240dp"
+        android:layout_width="match_parent"
+        android:maxWidth="240dp"
+        android:layout_margin="4dp"
         android:focusable="true"
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto">
