+++ app/src/main/res/layout/fragment_artists.xml
@@ -3,12 +3,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context=".ui.main.pages.ArtistsFragment">
-
-    <!-- TODO: Update blank fragment layout -->
-    <TextView
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:text="@string/hello_blank_fragment"/>
+        android:id="@+id/artistsFragment"
+             tools:context=".ui.mylibrary.pages.ArtistsFragment">
 
 </FrameLayout>
