+++ app/src/main/res/layout/fragment_home.xml
@@ -3,11 +3,8 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.main.home.HomeFragment">
+        android:id="@+id/homeFragment"
+        tools:context=".ui.home.HomeFragment">
 
-    <TextView
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:text="Home Fragment" />
 
 </FrameLayout>
