--- app/src/main/res/layout/fragment_menu.xml
@@ -1,54 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.appcompat.widget.LinearLayoutCompat
-        xmlns:android="http://schemas.android.com/apk/res/android"
-             xmlns:tools="http://schemas.android.com/tools"
-        android:orientation="vertical"
-             android:layout_width="match_parent"
-             android:layout_height="match_parent" xmlns:app="http://schemas.android.com/apk/res-auto"
-             tools:context=".ui.main.MenuFragment">
-
-    <!--
-    <com.google.android.material.tabs.TabLayout
-            style="@style/Widget.MaterialComponents.TabLayout"
-            android:id="@+id/tabLayout"
-            app:tabMode="scrollable"
-            app:tabIndicator="@drawable/tabs_rounded_line_indicator"
-            app:tabIndicatorFullWidth="false"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-
-        <com.google.android.material.tabs.TabItem
-                android:text="@string/artists"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"/>
-
-        <com.google.android.material.tabs.TabItem
-                android:text="@string/albums"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"/>
-
-        <com.google.android.material.tabs.TabItem
-                android:text="@string/songs"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"/>
-
-        <com.google.android.material.tabs.TabItem
-                android:text="@string/categories"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"/>
-
-        <com.google.android.material.tabs.TabItem
-                android:text="@string/playlists"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"/>
-    </com.google.android.material.tabs.TabLayout>
-
-    <androidx.viewpager.widget.ViewPager
-            android:id="@+id/viewPager"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-    </androidx.viewpager.widget.ViewPager>
-
-    -->
-
-</androidx.appcompat.widget.LinearLayoutCompat>
