+++ app/src/main/res/layout/fragment_playlists.xml
@@ -3,6 +3,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context=".ui.main.playlists.PlaylistsFragment">
+        android:id="@+id/playlistsFragment"
+             tools:context=".ui.mylibrary.pages.PlaylistsFragment">
 
 </FrameLayout>
