+++ app/src/main/res/layout/fragment_settings.xml
@@ -3,12 +3,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
+        android:id="@+id/settingsFragment"
              tools:context=".ui.preferences.SettingsFragment">
-
-    <!-- TODO: Update blank fragment layout -->
-    <TextView
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:text="@string/hello_blank_fragment"/>
 
 </FrameLayout>
