+++ app/src/main/res/layout/fragment_songs.xml
@@ -3,7 +3,8 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context=".ui.main.pages.SongsFragment">
+        android:id="@+id/songsFragment"
+             tools:context=".ui.mylibrary.pages.SongsFragment">
 
     <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/recyclerView"
