+++ app/src/main/res/menu/main_menu.xml
@@ -1,11 +1,21 @@
 <?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android" xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item
-            android:icon="@drawable/ic_home_alt_24dp"
-            android:id="@id/homeFragment"
-            android:title="@string/home">
-    </item>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
 
+    <item
+            android:id="@id/homeFragment"
+            android:icon="@drawable/ic_home_alt_24dp"
+            android:title="@string/home" />
+
+    <item
+            android:id="@id/myLibraryFragment"
+            android:icon="@drawable/ic_outline_library_music_24"
+            android:title="@string/my_library" />
+
+    <item
+            android:id="@id/myProfileFragment"
+            android:icon="@drawable/ic_outline_person_outline_24"
+            android:title="@string/my_profile" />
+    <!--
     <item
             android:icon="@drawable/ic_people_alt_24dp"
             android:id="@id/artistsFragment"
@@ -23,4 +33,5 @@
             android:id="@id/playlistsFragment"
             android:title="@string/playlists">
     </item>
+    -->
 </menu>
