+++ app/src/main/res/navigation/nav_graph.xml
@@ -8,22 +8,32 @@
               android:name="com.android.sergiobelda.gramophone.ui.album.AlbumDetailFragment"
               android:label="fragment_album_detail" tools:layout="@layout/fragment_album_detail"/>
     <fragment android:id="@+id/albumsFragment"
-              android:name="com.android.sergiobelda.gramophone.ui.main.pages.AlbumsFragment"
+              android:name="com.android.sergiobelda.gramophone.ui.mylibrary.pages.AlbumsFragment"
               android:label="fragment_albums" tools:layout="@layout/fragment_albums">
         <action android:id="@+id/showAlbum"
                 app:destination="@id/albumDetailFragment"/>
     </fragment>
     <fragment android:id="@+id/artistsFragment"
-              android:name="com.android.sergiobelda.gramophone.ui.main.pages.ArtistsFragment"
+              android:name="com.android.sergiobelda.gramophone.ui.mylibrary.pages.ArtistsFragment"
               android:label="fragment_artists" tools:layout="@layout/fragment_artists"/>
     <fragment
             android:id="@+id/homeFragment"
-            android:name="com.android.sergiobelda.gramophone.ui.main.home.HomeFragment"
+            android:name="com.android.sergiobelda.gramophone.ui.home.HomeFragment"
             android:label="@string/app_name"
             tools:layout="@layout/fragment_home" />
     <fragment
             android:id="@+id/playlistsFragment"
-            android:name="com.android.sergiobelda.gramophone.ui.main.playlists.PlaylistsFragment"
+            android:name="com.android.sergiobelda.gramophone.ui.mylibrary.pages.PlaylistsFragment"
             android:label="fragment_playlists"
             tools:layout="@layout/fragment_playlists" />
+    <fragment
+            android:id="@+id/myLibraryFragment"
+            android:name="com.android.sergiobelda.gramophone.ui.mylibrary.MyLibraryFragment"
+            android:label="fragment_my_library"
+            tools:layout="@layout/fragment_my_library" />
+    <fragment
+            android:id="@+id/myProfileFragment"
+            android:name="com.android.sergiobelda.gramophone.ui.myprofile.MyProfileFragment"
+            android:label="fragment_profile"
+            tools:layout="@layout/fragment_my_profile" />
 </navigation>
