+++ app/src/main/res/values/strings.xml
@@ -14,4 +14,5 @@
     <string name="search">Buscar</string>
     <string name="my_library">Mi librería</string>
     <string name="home">Inicio</string>
+    <string name="my_profile">Mi perfil</string>
 </resources>
