+++ build.gradle
@@ -5,7 +5,7 @@ buildscript {
         appcompat_version = '1.1.0'
         ktx_version = '1.1.0'
         fragment_ktx_version = '1.2.0-rc01'
-        constraint_version = '2.0.0-beta3'
+        constraint_version = '2.0.0-beta2'
         preference_ktx_version = '1.1.0'
         palette_ktx_version = '1.0.0'
         material_version = '1.2.0-alpha01'
